+++ src/main/java/webserver/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package domain;
+package webserver.http;
 
 public enum HttpMethod {
     GET, POST
